feat: Split joiner in three distinct stages
